requirements.txt
discord.py==1.7.3
jishaku==2.1.0
